SPUI-321 > Replace bid-ask with base-prices thread (#127)
* Replace bidask with base-prices thread

* Remove bidAsk's related functionality

* Change accuracy type from double to int

+++ lib/app/shared/features/currency_buy/notifier/currency_buy_notifier.dart
@@ -121,7 +121,7 @@ class CurrencyBuyNotifier extends StateNotifier<CurrencyBuyState> {
     if (state.targetConversionPrice != null && state.inputValue.isNotEmpty) {
       final amount = double.parse(state.inputValue);
       final price = state.targetConversionPrice!;
-      final accuracy = currencyModel.accuracy.toInt();
+      final accuracy = currencyModel.accuracy;
       var conversion = 0.0;
 
       if (price != 0) {
